--- pyaptamer/pseaac/__init__.py
@@ -1,6 +1,0 @@
-"""The PSeAAC encoding algorithm"""
-
-from pyaptamer.pseaac._pseaac_aptanet import AptaNetPSeAAC
-from pyaptamer.pseaac._pseaac_general import PSeAAC
-
-__all__ = ["PSeAAC", "AptaNetPSeAAC"]
