+++ pyaptamer/trafos/pseaac/__init__.py
@@ -1,0 +1,6 @@
+"""The PSeAAC encoding algorithm"""
+
+from pyaptamer.trafos.pseaac._pseaac_aptanet import AptaNetPSeAAC
+from pyaptamer.trafos.pseaac._pseaac_general import PSeAAC
+
+__all__ = ["PSeAAC", "AptaNetPSeAAC"]
