+++ pyaptamer/trafos/pseaac/tests/_props.py
